refactor: Rename `sdk-js/src/lib/spanHooks`
