+++ sdk-js/src/lib/span-hooks/hookAwsSdk.js
