+++ sdk-js/src/lib/span-hooks/hookHttp.js
